pyproject.toml
[project]
name = "agent-kms"
version = "0.1.0"
description = "Generic AI agent knowledge management system: Qdrant-backed retrieval + Claude Code Stop-hook lesson extraction. Extracted from a working migration project, designed for any agent that benefits from cross-session memory."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Toshihiro Yamaguchi" },
]
keywords = ["rag", "mcp", "claude-code", "qdrant", "knowledge-management", "agent", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pydantic>=2.7,<3.0",
    "qdrant-client>=1.10.0",
    "sentence-transformers>=3.0",
    "fastmcp>=0.2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
gemini = ["google-genai>=0.3"]
anthropic = ["anthropic>=0.40.0"]
all = ["google-genai>=0.3", "anthropic>=0.40.0", "fastembed>=0.4"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.6"]

[project.scripts]
agent-kms = "agent_kms.cli:main"

[project.urls]
Homepage = "https://github.com/spin-glass/agent-kms"
Issues = "https://github.com/spin-glass/agent-kms/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_kms"]
artifacts = ["agent_kms/presets/**"]

[tool.hatch.build.targets.sdist]
include = [
    "agent_kms",
    "scripts",
    "compose.yaml",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]
